Improve the webhook
Make it more modular and add rendered html pages
