+++ wh/app/app.rb
@@ -11,92 +11,13 @@ require 'sinatra'
 require 'yaml'
 
 # VARs
-CONFIG = ENV['API_CONFIG']
+CONFIG = ENV['API_CONFIG'] || 'config.yml'
 
-# Logging
-def logger
-  @logger ||= Logger.new(STDOUT)
-end
-
-# Configuration
-def config
-  @config ||= YAML.load_file(CONFIG)
-end
-
-# Filter running containers by service label
-def container(label, value)
-  @container = Docker::Container.all(
-    all: false,
-    filters: {
-      label: ["#{label}=#{value}"]
-    }.to_json
-  )
-rescue Excon::Error::Socket
-  logger.warn "Could not connect to docker daemon!"
-  return nil
-end
-
-# Deploy R10K in a separate thread (look for a container labeled with r10k)
-def deploy_r10k
-  Thread.new do
-    begin
-      logger.info 'Deploying R10K environment'
-      stdout, stderr, status = container('r10k', 'true').first
-        .exec(['r10k', 'deploy', 'environment', '--puppetfile'])
-      if status == 0
-        logger.info 'Deployment completed'
-      else
-        raise stdout.join(', ') unless stdout.empty?
-        raise stderr.join(', ') unless stderr.empty?
-      end
-    rescue => error
-      logger.warn "Deployment failed (#{error})!"
-    end
-  end
-end
-
-# Only allow authorized requests
-def protected!
-  return if authorized?
-  response['WWW-Authenticate'] = %(Basic realm="Restricted Area")
-  halt 401, 'Not authorized'
-end
-
-# Check credentials
-def authorized?
-  @auth ||= Rack::Auth::Basic::Request.new(request.env)
-  @auth.provided? && @auth.basic? && @auth.credentials && @auth.credentials == [config['user'], config['pass']]
-end
-
-# Verify GitHub signature
-def verify_github_signature(payload_body)
-  signature = 'sha1=' + OpenSSL::HMAC.hexdigest(OpenSSL::Digest::SHA1.new, config['github_secret'], payload_body)
-  return halt 500, 'Signatures did not match!' unless Rack::Utils.secure_compare(signature, request.env['HTTP_X_HUB_SIGNATURE'])
-end
-
-# Verify Travis signature
-def verify_travis_signature(payload)
-  signature = request.env['HTTP_SIGNATURE']
-  pkey      = OpenSSL::PKey::RSA.new(travis_public_key)
-
-  return halt 500, 'Signatures did not match!' unless pkey.verify(OpenSSL::Digest::SHA1.new, Base64.decode64(signature), payload.to_json)
-end
-
-# Get Travis public key
-def travis_public_key
-  return @travis_public_key if defined? @travis_public_key
-
-  @conn = Faraday.new(url: 'https://api.travis-ci.org') do |faraday|
-    faraday.adapter Faraday.default_adapter
-  end
-  response = @conn.get '/config'
-  JSON.parse(response.body)['config']['notifications']['webhook']['public_key']
-end
-
-# Initial deployment
-def initial_deployment
-  logger.info 'Start initial deployment'
-  deploy_r10k
+# Include helper methods
+require './lib/helpers'
+extend Helpers
+helpers do
+  include Helpers
 end
 
 # Configure Sinatra
@@ -114,8 +35,26 @@ $stdout.sync = true
 initial_deployment unless settings.test?
 
 # Home
-get '/hooks/?' do
-  'Nothing here! Yet!'
+get '/' do
+  erb :index
+end
+
+# Status
+get '/status' do
+  return 200,  {:status => :success, :message => 'running' }.to_json
+end
+
+# Environment info
+get '/env' do
+  protected!
+  if params[:json] == 'yes'
+    content_type :json
+    ENV.to_h.to_json
+  else
+    'Environment (as <a href="/env?json=yes">JSON</a>):<ul>' +
+      ENV.each.map { |key, value| "<li><b>#{key}:</b> #{value}</li>" }
+      .join + '</ul>'
+  end
 end
 
 # Travis webhook
@@ -154,7 +93,8 @@ post '/hooks/slack' do
     logger.info "Authorized request received from slacker @#{user} " \
                 "on channel ##{channel}"
   else
-    logger.warn "Unauthorized token received from slacker @#{user}"
+    logger.warn "Unauthorized token received from slacker @#{user}" \
+                "on channel ##{channel}"
   end
 
   case command
@@ -170,22 +110,4 @@ post '/hooks/slack' do
   else
     "Unknown command '#{command}' :cry:"
   end
-end
-
-# Environment info
-get '/env' do
-  protected!
-  if params[:json] == 'yes'
-    content_type :json
-    ENV.to_h.to_json
-  else
-    'Environment (as <a href="/env?json=yes">JSON</a>):<ul>' +
-      ENV.each.map { |key, value| "<li><b>#{key}:</b> #{value}</li>" }
-      .join + '</ul>'
-  end
-end
-
-# Status
-get '/status' do
-  return 200,  {:status => :success, :message => 'running' }.to_json
 end
