+++ wh/app/views/index.erb
@@ -1,0 +1,6 @@
+<div id="centered">
+  <img src="logo.png" width="300px" height="300px" alt="VGH" />
+  <footer>
+    <small>&copy; 2017, <a href="http://vladgh.com">VladGh.com</a>. All rights reserved.</small>
+  </footer>
+</div>
